moved AssetPrice domain model to model/assets package
